// src/contract-apis/types/StrandsLyraVault.ts
export declare namespace Vault {
  export type VaultParamsStruct = {
    decimals: PromiseOrValue<BigNumberish>;
    cap: PromiseOrValue<BigNumberish>;
    asset: PromiseOrValue<string>;
  };

  export type VaultParamsStructOutput = [number, BigNumber, string] & {
    decimals: number;
    cap: BigNumber;
    asset: string;
  };
}